move data type workspace context token
