+++ src/packages/core/data-type/index.ts
@@ -1,6 +1,8 @@
 import './components/index.js';
 
+export * from './components/index.js';
 export * from './entity.js';
 export * from './repository/index.js';
+export * from './workspace/index.js';
 
 export type { UmbDataTypeDetailModel } from './types.js';
